requirements.txt
# Core NLP and embedding
sentence-transformers>=2.2.2

# LangChain ecosystem
langchain
langchain-community
langchain-openai
langchain-pinecone
langchain-experimental
langchain-huggingface
# Web framework
flask

# PDF parsing
pypdf

# Environment variables
python-dotenv

# Pinecone (latest without deprecated plugin support)
pinecone-client<6.0.0
# PyTorch (CUDA 12.8 build)
torch==2.2.2+cu128
torchvision==0.17.2+cu128
torchaudio==2.2.2+cu128
--extra-index-url https://download.pytorch.org/whl/cu128

# Editable install (for local module)
-e .

bitsandbytes
accelerate>=0.26.0
chainlit